+++ src/FluxJsonFileSettingsStorage.mjs
@@ -1,15 +1,15 @@
-/** @typedef {import("./FluxFileSettingsStorage.mjs").FluxFileSettingsStorage} FluxFileSettingsStorage */
+/** @typedef {import("./SettingsStorage.mjs").SettingsStorage} SettingsStorage */
 
 export class FluxJsonFileSettingsStorage {
     /**
      * @param {string} file_path
-     * @returns {Promise<FluxFileSettingsStorage>}
+     * @returns {Promise<SettingsStorage>}
      */
     static async new(file_path) {
         return (await import("./FluxFileSettingsStorage.mjs")).FluxFileSettingsStorage.new(
             file_path,
-            async string => JSON.parse(string),
-            async settings => JSON.stringify(settings)
+            async settings => JSON.stringify(settings),
+            async settings => JSON.parse(settings)
         );
     }